cleaned up the css so far
