+++ app/(auth)/login.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { StyleSheet, Image, TextInput, TouchableOpacity, Alert, ActivityIndicator } from 'react-native';
 import { useRouter } from 'expo-router';
-
+import { styles } from '../css/login';
 import { ThemedText } from '@/components/ThemedText';
 import { ThemedView } from '@/components/ThemedView';
 import { useAuth } from '../context/AuthContext';
@@ -129,102 +129,3 @@ export default function LoginScreen(): JSX.Element {
     </ThemedView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#f8fafc',
-
-    justifyContent: 'center',
-    padding: 20,
-  },
-  logoContainer: {
-    alignItems: 'center',
-    marginBottom: 20,
-    backgroundColor: '#f8fafc'
-
-  },
-  logo: {
-    width: 200,
-    height: 200,
-    resizeMode: 'contain',
-    backgroundColor: '#f8fafc',
-
-  },
-  card: {
-    backgroundColor: 'white',
-    borderRadius: 16,
-    padding: 24,
-    width: '100%',
-    maxWidth: 500,
-    alignSelf: 'center',
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.06,
-    shadowRadius: 6,
-    elevation: 3,
-  },
-  title: {
-    fontSize: 28,
-    fontWeight: 'bold',
-    color: '#1e40af',
-    textAlign: 'center',
-    marginBottom: 4,
-  },
-  subtitle: {
-    fontSize: 16,
-    color: '#64748b',
-    textAlign: 'center',
-    marginBottom: 24,
-  },
-  inputContainer: {
-    backgroundColor: 'white',
-    width: '100%'
-  },
-  label: {
-    marginBottom: 8,
-    fontWeight: '600',
-    color: '#475569',
-    fontSize: 15,
-  },
-  input: {
-    height: 50,
-    borderWidth: 1,
-    borderColor: '#e2e8f0',
-    borderRadius: 8,
-    marginBottom: 16,
-    paddingHorizontal: 12,
-    fontSize: 16,
-    backgroundColor: '#f8fafc',
-  },
-  button: {
-    backgroundColor: '#3b82f6',
-    height: 50,
-    borderRadius: 8,
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginBottom: 12,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 1 },
-    shadowOpacity: 0.05,
-    shadowRadius: 2,
-    elevation: 1,
-  },
-  greenButton: {
-    backgroundColor: '#22c55e',
-  },
-  altButton: {
-    backgroundColor: '#64748b',
-  },
-  buttonText: {
-    color: 'white',
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  helperText: {
-    marginTop: 12,
-    textAlign: 'center',
-    color: '#64748b',
-    fontSize: 14,
-  },
-});
