+++ app/(tabs)/Vehicles.tsx
@@ -1,4 +1,6 @@
 import React, { useState } from 'react';
+import styles from '../css/vehicles';
+
 import {
   View,
   Text,
@@ -363,262 +365,3 @@ export default function RoadTripPlannerApp() {
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#f8fafc',
-  },
-  header: {
-    marginTop: 60,
-    alignItems: 'center',
-    padding: 24,
-    backgroundColor: 'white',
-    borderBottomWidth: 1,
-    borderBottomColor: '#f1f5f9',
-  },
-  title: {
-    fontSize: 28,
-    fontWeight: 'bold',
-    color: '#1e40af',
-  },
-  subtitle: {
-    fontSize: 16,
-    color: '#64748b',
-    textAlign: 'center',
-    marginTop: 8,
-  },
-  main: {
-    padding: 16,
-  },
-  sectionSeparator: {
-    height: 16,
-  },
-  sectionHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 16,
-    backgroundColor: 'white',
-    padding: 20,
-    borderRadius: 16,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.06,
-    shadowRadius: 6,
-    elevation: 3,
-  },
-  sectionTitle: {
-    fontSize: 22,
-    fontWeight: '700',
-    color: '#0f172a',
-  },
-  button: {
-    backgroundColor: '#3b82f6',
-    paddingVertical: 10,
-    paddingHorizontal: 16,
-    borderRadius: 8,
-    alignItems: 'center',
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 1 },
-    shadowOpacity: 0.05,
-    shadowRadius: 2,
-    elevation: 1,
-  },
-  buttonText: {
-    color: 'white',
-    fontWeight: '600',
-    fontSize: 15,
-  },
-
-  vehicleItem: {
-    flexDirection: 'row',
-    backgroundColor: 'white',
-    padding: 16,
-    borderRadius: 12,
-    marginBottom: 8,
-    borderLeftWidth: 4,
-    borderLeftColor: '#3b82f6',
-    alignItems: 'center',
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 1 },
-    shadowOpacity: 0.05,
-    shadowRadius: 2,
-    elevation: 1,
-  },
-  vehicleImage: {
-    width: 80,
-    height: 60,
-    borderRadius: 8,
-    marginRight: 16,
-    backgroundColor: '#e2e8f0',
-  },
-  vehicleDetails: {
-    flex: 1,
-  },
-  vehicleName: {
-    fontSize: 18,
-    fontWeight: '700',
-    color: '#0f172a',
-    marginBottom: 4,
-  },
-  vehicleSpecs: {
-    color: '#475569',
-    marginBottom: 2,
-    fontSize: 14,
-  },
-
-  collaboratorItem: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    padding: 12,
-    backgroundColor: 'white',
-    borderRadius: 12,
-    marginBottom: 8,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 1 },
-    shadowOpacity: 0.05,
-    shadowRadius: 2,
-    elevation: 1,
-  },
-  avatar: {
-    width: 40,
-    height: 40,
-    borderRadius: 20,
-    marginRight: 12,
-    backgroundColor: '#e2e8f0',
-  },
-  collaboratorInfo: {
-    flex: 1,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  collaboratorName: {
-    fontSize: 16,
-    fontWeight: '600',
-    color: '#0f172a',
-  },
-  roleBadge: {
-    paddingVertical: 4,
-    paddingHorizontal: 10,
-    borderRadius: 16,
-  },
-  roleText: {
-    color: 'white',
-    fontWeight: '600',
-    fontSize: 12,
-  },
-
-  emptyState: {
-    alignItems: 'center',
-    padding: 24,
-    backgroundColor: 'white',
-    borderRadius: 12,
-    marginBottom: 8,
-  },
-  emptyStateText: {
-    fontSize: 18,
-    fontWeight: '600',
-    color: '#64748b',
-    marginBottom: 8,
-  },
-  emptyStateSubtext: {
-    color: '#94a3b8',
-    textAlign: 'center',
-  },
-
-  modalOverlay: {
-    flex: 1,
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 20,
-  },
-  modalContent: {
-    backgroundColor: 'white',
-    borderRadius: 16,
-    padding: 24,
-    width: '100%',
-    maxWidth: 500,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 4 },
-    shadowOpacity: 0.1,
-    shadowRadius: 8,
-    elevation: 5,
-  },
-  modalTitle: {
-    fontSize: 22,
-    fontWeight: '700',
-    marginBottom: 20,
-    color: '#0f172a',
-    textAlign: 'center',
-  },
-
-  form: {
-    gap: 16,
-  },
-  formGroup: {
-    marginBottom: 12,
-  },
-  formRow: {
-    flexDirection: 'row',
-    marginBottom: 12,
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: '#e2e8f0',
-    borderRadius: 8,
-    padding: 12,
-    fontSize: 16,
-    backgroundColor: '#f8fafc',
-  },
-  label: {
-    marginBottom: 8,
-    fontWeight: '600',
-    color: '#475569',
-    fontSize: 15,
-  },
-  modalButtons: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginTop: 24,
-  },
-  saveButton: {
-    backgroundColor: '#22c55e',
-    flex: 1,
-  },
-  cancelButton: {
-    backgroundColor: '#f1f5f9',
-    marginRight: 12,
-    flex: 1,
-  },
-  cancelButtonText: {
-    color: '#64748b',
-    fontWeight: '600',
-    fontSize: 15,
-  },
-
-  roleSelector: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  roleOption: {
-    flex: 1,
-    paddingVertical: 12,
-    paddingHorizontal: 8,
-    alignItems: 'center',
-    borderWidth: 2,
-    borderColor: '#e2e8f0',
-    borderRadius: 8,
-    marginHorizontal: 4,
-  },
-  selectedRoleOption: {
-    backgroundColor: '#f8fafc',
-    borderWidth: 2,
-  },
-  roleOptionText: {
-    fontWeight: '600',
-    color: '#64748b',
-  },
-});
